% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FASt_CDF.R
\name{FASt_cdf}
\alias{FASt_cdf}
\title{FASt cdf algorithm}
\usage{
FASt_cdf(
  x,
  mu = rep(0, length(x)),
  Sigma = diag(length(x)),
  log.p = FALSE,
  eps = 2,
  tolerance = 1e-05,
  algorithm = c("B", "A"),
  method = c("chol", "eigen")
)
}
\arguments{
\item{x}{the vector of quantiles of interest.}

\item{mu}{mean vector of the multivariate Gaussian of interest.}

\item{Sigma}{variance-Covariance matrix of the multivariate Gaussian of interest.}

\item{log.p}{logical. If \code{TRUE}, the log-cumulative probability is returned. Default is \code{FALSE}.}

\item{eps}{scalar, rescaling coefficient for the smallest eigenvalue (see Fasano and Denti, 2025+ for details).}

\item{tolerance}{scalar, a small positive value specifying the stopping criterion.}

\item{algorithm}{Can be either c("A","B"),}

\item{method}{Can be either c("chol","eigen")}
}
\value{
The estimated cumulative (log-) probability.
}
\description{
This function computes the value of a generic cumulative distribution function of a multivariate Gaussian
exploiting its connection with the marginal likelihood of a tailored dual Bayesian probit model.
}
\details{
Let us consider, without loss of generality, an \eqn{m}-dimensional multivariate Gaussian random variable with mean 0 and covariance matrix \eqn{\Sigma}. Then, for \eqn{u = (u_1,\ldots,u_m)^\top\in \mathbb{R}^m}, its cumulative distribution function is defined as

\deqn{\Phi_m(u;\Sigma) = \int_{{-\infty}}^{u} \int_{C_u}\dfrac{1}{\sqrt{(2\pi)^m|\Sigma|}} \exp \left\{-\dfrac{1}{2} w^\top \Sigma^{-1} w \right\} d w,}
where \eqn{C_u=\left\{w=(w_1,\ldots,w_m)^\top\in \mathbb{R}^m\colon w_i\le u_i,\ \forall i=1,\ldots,m \right\}}.


Given any positive definite \eqn{m\times m} covariance matrix \eqn{\Sigma}, call \eqn{\lambda_{m}} its smallest eigenvalue and define \eqn{\tilde{\Sigma}=\Sigma-\epsilon\lambda_{m}I_m}, where \eqn{\epsilon\in (0,1)}.
Consider now any factorization of \eqn{\tilde{\Sigma}} of the form \eqn{\tilde{\Sigma}=P\tilde\Lambda P^\top}, where \eqn{P} and \eqn{\tilde\Lambda} are \eqn{m\times m} matrices
with the former invertible and the latter diagonal and positive definite. Then, for any \eqn{u\in\mathbb{R}^m}, \eqn{\Phi_m(u;\Sigma)} \strong{equals the marginal likelihood of a dual Bayesian probit model}.

Specifically, consider a generic observation vector \eqn{y\in\{0,1\}^n} and design matrix \eqn{X\in\mathbb{R}^{n\times p}}, with \eqn{i}-th row denoted with \eqn{x_i^\top}, indicating the covariate vector for observation \eqn{i}.
Then, this model can be defined as
\deqn{y_i\mid \beta\overset{ind}{\sim} \text{Bern}\left\{\Phi_1((2y_i-1)x_i^\top\beta;1)\right\}, \quad i=1,\ldots,n,\quad   \beta\sim\mathcal{N}_p\left({\xi},\Omega\right),}
where, in particular, \eqn{n=p=m}, \eqn{y_i=1} for all \eqn{i=1,\ldots,m,} and
\deqn{{\xi} =\left(\epsilon \lambda_{m}\right)^{-1/2}P^{-1}u,\quad \Omega= \left(\epsilon \lambda_{m}\right)^{-1}\tilde\Lambda,\quad X = P.}

If the Gaussian of interest is not centered, one can assume \eqn{u = x-\mu}.

The function implements several algorithms to estimate this dual Bayesian probit model via Expectation Propagation, and then computing its marginal likelihood.
}
\examples{
FASt_cdf(c(0,0,0), log = FALSE)
}
\references{
Fasano, A., and Denti, F. (2025+). Multivariate Gaussian cumulative distribution functions as the marginal likelihood of their dual Bayesian probit models.
\emph{ArXiv}
}
